+++ src/pages/PostIdMessage/PostMessage.jsx
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom';
 import { useNavigate } from 'react-router-dom';
 
 import postRecipientMessage from '@/apis/post/postRecipientsMessage';
-import Button from '@/components/common/Buttons/Button';
+import Button from '@/components/common/button/Button';
 import DropDown from '@/components/common/DropDown/DropDown';
 import GlobalHeader from '@/components/common/header/GlobalHeader';
 import Input from '@/components/common/Input/Input';
@@ -110,9 +110,10 @@ function PostMessage() {
 				<FontSelector onSelectFont={handleFontChange} />
 				<Link to={`/post/${recipientId}`} className='link'>
 					<Button
-						size='basic'
+						variant='primary'
+						size='full'
 						disabled={isButtonDisabled}
-						content='생성하기'
+						label='생성하기'
 						onClick={handleCreateMessage}
 						type='button'
 					/>
