+++ src/pages/post/Select.scss
@@ -1,4 +1,3 @@
-
 @use '@/assets/utilities.scss' as *;
 $select-label-margin: 4rem;
 
@@ -12,11 +11,6 @@ $select-label-margin: 4rem;
 	display: flex;
 	flex-direction: row;
 	margin: 2.4rem 0;
-}
-
-.toggleBox > .button {
-	width: 9rem;
-	margin: 0;
 }
 
 .selectLabel {
@@ -48,10 +42,6 @@ $select-label-margin: 4rem;
 
 .checked svg {
 	display: block;
-}
-
-.postForm > .button {
-	margin-top: 10rem;
 }
 
 .checkboxInput {
